Create README.md
Sushi Mama is a local teriyaki grill serving steaks and seafood, plus sushi and a traditional Japanese menu. They offer dine-in, takeout, and delivery. Here is an updated website offering a view of their menu and online ordering options, as well as an online contact form for comments and feedback.